move implementations of dataset and group forms to separate plugins
resolve issue with unrecognized templates because of method name conflicts
